refactor(core): move I18n util to shared package
