--- core/src/main/java/net/lapidist/colony/i18n/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Translation system backed by resource bundles.
- * Provides the {@link I18n} utility to access localized strings.
- */
-package net.lapidist.colony.i18n;
