+++ core/src/main/java/net/lapidist/colony/util/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Utility classes shared across modules.
+ * Provides the {@link I18n} class to access localized strings.
+ */
+package net.lapidist.colony.util;
