+++ docs/i18n.md
@@ -6,7 +6,7 @@ All user-facing text is stored in property files and retrieved at runtime throug
 
 ## I18n Utility
 
-The class `net.lapidist.colony.i18n.I18n` exposes two simple methods:
+The class `net.lapidist.colony.util.I18n` exposes two simple methods:
 
 - `I18n.get("key")` – return the translated string for `key`. Missing keys are
   returned as `!key!` so they are easy to spot during development.
